fix small bug causes from previous change
