+++ src/Command/Git/MergeDown.php
@@ -8,8 +8,9 @@ use Tomato\Command\AbstractCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Tomato\Command\AbstractGitCommand;
 
-class MergeDown extends AbstractCommand
+class MergeDown extends AbstractGitCommand
 {
     protected function configure()
     {
