+++ src/Command/Git/Projects.php
@@ -9,8 +9,9 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Tomato\Command\AbstractCommand;
+use Tomato\Command\AbstractGitCommand;
 
-class Projects extends AbstractCommand
+class Projects extends AbstractGitCommand
 {
     /**
      * @var Client
